add pageable and infinite update for list and normal update for map

+++ app/src/main/java/ru/hse/equeue/model/Queue.kt
@@ -1,7 +1,7 @@
 package ru.hse.equeue.model
 
 data class Queue(
-    val id: Double,
+    val id: Long,
     val photoUrl: String,
     val name: String,
     val x: Double,
